refactor: extract model selection from AgentWorkspace
